+++ dashboard.php
@@ -183,8 +183,8 @@ $utilisateur = $_SESSION['auth']->id;
 																</div>
 															</th>
 															<th class="min-w-150px">Nom du produit</th>
-															<th class="min-w-140px">Prix</th>
-															<th class="min-w-120px">Etats</th>
+															<th class="min-w-140px">Prix<br>Type de remboursement</th>
+															<th class="min-w-120px">Etats<br>Date de review prevu</th>
 															<th class="min-w-100px text-end">Actions</th>
 														</tr>
 													</thead>
@@ -192,6 +192,7 @@ $utilisateur = $_SESSION['auth']->id;
 													<?php
 													$db = App::getDatabase()->query("SELECT * FROM products WHERE username=$utilisateur ORDER BY id DESC LIMIT 6");
 													while ($info = $db->fetch(PDO::FETCH_ASSOC)) {
+													$review_date = date_create($info['review_date']);
 													echo
 													//begin::Table body
 													'<tbody>
@@ -219,7 +220,8 @@ $utilisateur = $_SESSION['auth']->id;
 															<td class="text-end">
 																<div class="d-flex flex-column w-100 me-2">
 																	<div class="d-flex flex-stack mb-2">
-																		<span class="text-muted me-2 fs-7 fw-bold">'. $info['condition'] .'</span>
+																		<span class="text-muted me-2 fs-7 fw-bold">'. $info['condition'] .'test</span>
+																		<span class="text-muted me-2 fs-7 fw-bold">'. date_format($review_date, 'd/m') .'</span>
 																	</div>
 																	<div class="progress h-6px w-100">
 																		<div class="progress-bar bg-primary" role="progressbar" style="width: 50%" aria-valuenow="50" aria-valuemin="0" aria-valuemax="100"></div>