+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Passage de expectancy, resale, refund_type dans la base de donnée en NULL
 - Passage des checkbox pour le type de remboursement en radio
 - Ajout automatique de l'avis dans l'attente du vendeur
+- Modification du dashboard pour affiché le type de remboursement, l'etat de la commande et la date prévu de l'ajout de la review
 
 ## [1.8] - 2021-08-21
 ### Changed